requirements-dev.txt
black>=21.9b0
eth-brownie>=1.16.4
click>=8.0.1